Create a module for DB
