+++ app/build.gradle
@@ -49,6 +49,8 @@ android {
 
 dependencies {
     implementation project(":core:common")
+    implementation project(":core:database")
+
     implementation "androidx.media:media:$media_api_version"
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:1.8.0"
     implementation "androidx.appcompat:appcompat:$appcompat_version"
@@ -80,11 +82,6 @@ dependencies {
     kapt "androidx.hilt:hilt-compiler:$androidx_hilt_compiler_version"
     implementation "com.github.shuhart:stepview:$stepview_version"
     implementation "com.github.kenglxn.QRGen:android:$qr_gen_version"
-
-    // Room
-    implementation "androidx.room:room-runtime:$room_version"
-    implementation "androidx.room:room-ktx:$room_version"
-    kapt "androidx.room:room-compiler:$room_version"
 
     implementation("com.google.api-client:google-api-client-android:$google_api_client_android_version") {
         exclude group: 'org.apache.httpcomponents'
