+++ app/src/main/java/com/yurii/youtubemusic/di/AppModule.kt
@@ -1,10 +1,8 @@
 package com.yurii.youtubemusic.di
 
 import android.content.Context
-import androidx.room.Room
 import androidx.work.WorkManager
 import com.github.kiulian.downloader.YoutubeDownloader
-import com.yurii.youtubemusic.db.DataBase
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -22,20 +20,6 @@ annotation class MainScope
 @Module
 @InstallIn(SingletonComponent::class)
 class AppModule {
-
-    @Provides
-    @Singleton
-    fun provideDataBase(@ApplicationContext context: Context) =
-        Room.databaseBuilder(context, DataBase::class.java, "database").build()
-
-    @Provides
-    fun provideMediaItemDao(dataBase: DataBase) = dataBase.mediaItemDao()
-
-    @Provides
-    fun providePlaylistDao(dataBase: DataBase) = dataBase.playlistDao()
-
-    @Provides
-    fun provideYouTubePlaylistSynchronizationDao(dataBase: DataBase) = dataBase.playlistSyncBindDao()
 
     @Provides
     @Singleton
