+++ core/database/src/main/java/com/youtubemusic/core/database/DatabaseModule.kt
@@ -1,0 +1,19 @@
+package com.youtubemusic.core.database
+
+import android.content.Context
+import androidx.room.Room
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.android.qualifiers.ApplicationContext
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+object DatabaseModule {
+    @Provides
+    @Singleton
+    fun provideDataBase(@ApplicationContext context: Context) =
+        Room.databaseBuilder(context, DataBase::class.java, "database").build()
+}
