+++ core/database/src/main/java/com/youtubemusic/core/database/models/PlaylistEntity.kt
@@ -1,0 +1,10 @@
+package com.youtubemusic.core.database.models
+
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+@Entity(tableName = "playlists")
+data class PlaylistEntity(
+    @PrimaryKey(autoGenerate = true) val playlistId: Long = 0,
+    val name: String
+)
