+++ settings.gradle
@@ -17,3 +17,4 @@ dependencyResolutionManagement {
 include ':app'
 rootProject.name='YouTubeMusic'
 include ':core:common'
+include ':core:database'
